Added some parsing functionality
Housing module has some internal functions now. One parses the PDF
and returns the pages of each PDF, split by the "Page x of y" string
as an indicator (hopefully there is a better way of doing this, as
well as avoiding a hard-storage version of the PDF).

The other function currently parses each page and adds each uni to
a group ID key within a dictionary.

Changes to be committed:
	modified:   .gitignore
	new file:   housing.py
	new file:   requirements.txt