Added map of UNI -> group_id, as well as ordered list of group_ids
Group IDs are essentially already sorted in the housing PDF,
so may as well take advantage of this functionality while parsing
the PDF to obtain a nice sorted list of housing groups.

Also added functionality to update a map of UNI -> group ID, in
addition to the map of group ID -> group UNIs.

Changes to be committed:
	modified:   housing.py