Added full functionality for parsing data
Public function in housing now returns a Housing object that
contains the group/student maps, as well as ordered arrays containing
group_ids and groups split by size.

Also moved PDF file out of the module and into another module that
simply contains the path; this will make it more modular when it comes
to changing the data per year.

Changes to be committed:
	modified:   .gitignore
	modified:   housing.py